Moved list-rooms file
